datasets/contextual_integrity/pyproject.toml
[project]
name = "contextual-integrity-eval"
version = "0.1.0"
description = "Evaluation framework for Contextual Integrity dataset"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["contextual_integrity_eval"]
